+++ bin/cli.js
@@ -1,25 +1,40 @@
 #!/usr/bin/env node
 const yargs = require('yargs/yargs')
 const { hideBin } = require('yargs/helpers')
-const extractor = require('../extractor')
+const extractor = require('../src/extractor')
+const path = require('path')
+const constants = require('../src/constants')
 
 yargs(hideBin(process.argv))
-    .command('run [locales]', 'extract localisable strings from the following input', (yargs) => {
+    .command('run [locales]', 'extract localisable strings from your project path', (yargs) => {
         return yargs.positional('locales', {
-            describe: 'Comma separated list of locales to generate into files'
+            describe: 'Comma separated list of locales'
         })
         .demandOption('locales')
         .options({
             'path': {
                 alias: 'p',
-                describe: 'Source path where start finding resources',
-                default: process.cwd()
+                describe: 'Source path where start finding for your project\'s code (can be relative)',
+                default: path.join(process.cwd(), 'resources', 'js')
+            },
+            'extensions': {
+                alias: 'e',
+                describe: 'Files extensions to match as input',
+                default: constants.extensions
+            },
+            'output': {
+                alias: 'o',
+                describe: 'Output path to place all generated locales',
+                default: path.join(process.cwd(), 'resources', 'lang')
+            },
+            'match': {
+                alias: 'm',
+                describe: 'Customize match regexp to search within input files',
+                default: constants.match
             }
         })
     }, (argv) => {
-        extractor(argv.locales, {
-            path: argv.path
-        })
+        extractor(argv.locales, argv)
     })
     .help()
     .parse()