+++ src/index.js
@@ -1,0 +1,4 @@
+const mix = require('laravel-mix');
+const I18n = require('./plugin');
+
+mix.extend('i18n', new I18n);